feat: conclui pratica07. Fecha #12
